Fix logging and move Node post processing actions to data_set_manager
